requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
crypto==1.4.1
cycler==0.10.0
idna==2.8
itchat==1.3.10
kiwisolver==1.0.1
matplotlib==3.0.2
Naked==0.1.31
numpy==1.16.1
Pillow==5.4.1
pycryptodome==3.7.3
pyparsing==2.3.1
pypng==0.0.19
PyQRCode==1.2.1
python-dateutil==2.7.5
PyYAML==3.13
requests==2.21.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1
wordcloud==1.5.0
